+++ resources/views/klasifikasi/index.blade.php
@@ -25,8 +25,7 @@
         </div>
         <div>
             <div class="col">
-                <button type="button" class="btn btn-sm btn-primary" data-toggle="modal"
-                    data-target="#tambahklasifikasi"><i class="fas fa-plus"></i>
+                <button type="button" class="btn btn-sm btn-primary" data-toggle="modal" data-target="#tambahklasifikasi"><i class="fas fa-plus"></i>
                     Tambah Data
                 </button>
             </div>
@@ -45,22 +44,18 @@
                         </tr>
                     </thead>
                     <tbody>
-                        <?php $no = 0;?>
+                        <?php $no = 0; ?>
                         @foreach($data_klasifikasi as $klasifikasi)
-                        <?php $no++ ;?>
+                        <?php $no++; ?>
                         <tr>
                             <td>{{$no}}</td>
                             <td>{{$klasifikasi->nama}}</td>
                             <td>{{$klasifikasi->kode}}</td>
                             <td>{{$klasifikasi->uraian}}</td>
                             <td>
-                                <a href="/klasifikasi/{{$klasifikasi->id}}/edit"
-                                    class="btn btn-primary btn-sm my-1 mr-sm-1"><i
-                                        class="nav-icon fas fa-pencil-alt"></i> Edit</a>
+                                <a href="/klasifikasi/{{$klasifikasi->id}}/edit" class="btn btn-primary btn-sm my-1 mr-sm-1"><i class="nav-icon fas fa-pencil-alt"></i> Edit</a>
                                 @if (auth()->user()->role == 'admin')
-                                <a href="/klasifikasi/{{$klasifikasi->id}}/delete"
-                                    class="btn btn-danger btn-sm my-1 mr-sm-1"
-                                    onclick="return confirm('Hapus Data ?')"><i class="nav-icon fas fa-trash"></i>
+                                <a href="/klasifikasi/{{$klasifikasi->id}}/delete" class="btn btn-danger btn-sm my-1 mr-sm-1" onclick="return confirm('Hapus Data ?')"><i class="nav-icon fas fa-trash"></i>
                                     Hapus</a>
                                 @endif
                             </td>
@@ -74,13 +69,11 @@
         </div>
 
         <!-- Modal Import -->
-        <div class="modal fade" id="importKlasifikasi" tabindex="-1" role="dialog" aria-labelledby="exampleModalLabel"
-            aria-hidden="true">
+        <div class="modal fade" id="importKlasifikasi" tabindex="-1" role="dialog" aria-labelledby="exampleModalLabel" aria-hidden="true">
             <div class="modal-dialog" role="document">
                 <div class="modal-content">
                     <div class="modal-header">
-                        <h5 class="modal-title" id="exampleModalLabel"><i
-                                class="nav-icon fas fa-layer-group my-1 btn-sm-1"></i> Import File Klasifikasi</h5>
+                        <h5 class="modal-title" id="exampleModalLabel"><i class="nav-icon fas fa-layer-group my-1 btn-sm-1"></i> Import File Klasifikasi</h5>
                         <button type="button" class="close" data-dismiss="modal" aria-label="Close">
                             <span aria-hidden="true">&times;</span>
                         </button>
@@ -90,8 +83,7 @@
 
                         {!! Form::file('data_klasifikasi.xls') !!} --}}
                         <div class="card-body">
-                            <form action="{{ url('klasifikasi.import') }}" method="POST" name="importform"
-                                enctype="multipart/form-data">
+                            <form action="{{ url('klasifikasi.import') }}" method="POST" name="importform" enctype="multipart/form-data">
                                 @csrf
                                 <input type="file" name="data_klasifikasi.xls" class="form-control">
                                 <br>
@@ -108,13 +100,11 @@
         </div>
 
         <!-- Modal Tambah -->
-        <div class="modal fade" id="tambahklasifikasi" tabindex="-1" role="dialog" aria-labelledby="exampleModalLabel"
-            aria-hidden="true">
+        <div class="modal fade" id="tambahklasifikasi" tabindex="-1" role="dialog" aria-labelledby="exampleModalLabel" aria-hidden="true">
             <div class="modal-dialog" role="document">
                 <div class="modal-content">
                     <div class="modal-header">
-                        <h5 class="modal-title" id="exampleModalLabel"><i
-                                class="nav-icon fas fa-layer-group my-1 btn-sm-1"></i> Tambah Data Klasifikasi</h5>
+                        <h5 class="modal-title" id="exampleModalLabel"><i class="nav-icon fas fa-layer-group my-1 btn-sm-1"></i> Tambah Data Klasifikasi</h5>
                         <button type="button" class="close" data-dismiss="modal" aria-label="Close">
                             <span aria-hidden="true">&times;</span>
                         </button>
@@ -125,14 +115,11 @@
                             <div class="row">
                                 <div class="col">
                                     <label for="nama">Nama</label>
-                                    <input name="nama" type="text" class="form-control bg-light" id="nama"
-                                        placeholder="Nama Klasifikasi" required>
+                                    <input name="nama" type="text" class="form-control bg-light" id="nama" placeholder="Nama Klasifikasi" required>
                                     <label for="kode">Kode</label>
-                                    <input name="kode" type="text" class="form-control bg-light" id="kode"
-                                        placeholder="Kode Klasifikasi" required>
+                                    <input name="kode" type="text" class="form-control bg-light" id="kode" placeholder="Kode Klasifikasi" required>
                                     <label for="uraian">Uraian</label>
-                                    <textarea name="uraian" class="form-control bg-light" id="uraian" rows="3"
-                                        placeholder="Uraian Klasifikasi" required></textarea>
+                                    <textarea name="uraian" class="form-control bg-light" id="uraian" rows="3" placeholder="Uraian Klasifikasi" required></textarea>
                                 </div>
                             </div>
                             <hr>
