+++ resources/views/tapel/index.blade.php
@@ -29,13 +29,12 @@
             </div>
         </div>
         <div>
-                <div class="col">
-                    <button type="button" class="btn btn-sm btn-primary" data-toggle="modal"
-                        data-target="#tambahTapel"><i class="fas fa-plus"></i>
-                        Tambah Data
-                    </button>
-                </div>
-                <br>
+            <div class="col">
+                <button type="button" class="btn btn-sm btn-primary" data-toggle="modal" data-target="#tambahTapel"><i class="fas fa-plus"></i>
+                    Tambah Data
+                </button>
+            </div>
+            <br>
         </div>
         <div class="row">
             <div class="row table-responsive">
@@ -50,24 +49,20 @@
                             </tr>
                         </thead>
                         <tbody>
-                            <?php $no = 0;?>
+                            <?php $no = 0; ?>
                             @foreach($data_tapel as $tapel)
-                            <?php $no++ ;?>
+                            <?php $no++; ?>
                             <tr>
                                 <td>{{$no}}</td>
                                 <td>{{$tapel->tahun}}</td>
                                 <td>{{$tapel->semester}}</td>
                                 <td>
-                                <a href="/tapel/{{$tapel->id}}/edit"
-                                    class="btn btn-primary btn-sm my-1 mr-sm-1"><i
-                                        class="nav-icon fas fa-pencil-alt"></i> Edit</a>
-                                @if (auth()->user()->role == 'admin')
-                                <a href="/tapel/{{$tapel->id}}/delete"
-                                    class="btn btn-danger btn-sm my-1 mr-sm-1"
-                                    onclick="return confirm('Hapus Data ?')"><i class="nav-icon fas fa-trash"></i>
-                                    Hapus</a>
-                                @endif
-                            </td>
+                                    <a href="/tapel/{{$tapel->id}}/edit" class="btn btn-primary btn-sm my-1 mr-sm-1"><i class="nav-icon fas fa-pencil-alt"></i> Edit</a>
+                                    @if (auth()->user()->role == 'admin')
+                                    <a href="/tapel/{{$tapel->id}}/delete" class="btn btn-danger btn-sm my-1 mr-sm-1" onclick="return confirm('Hapus Data ?')"><i class="nav-icon fas fa-trash"></i>
+                                        Hapus</a>
+                                    @endif
+                                </td>
                             </tr>
                             @endforeach
                         </tbody>
@@ -75,39 +70,36 @@
                 </div>
             </div>
             <!-- Modal Tambah -->
-        <div class="modal fade" id="tambahTapel" tabindex="-1" role="dialog" aria-labelledby="exampleModalLabel"
-            aria-hidden="true">
-            <div class="modal-dialog" role="document">
-                <div class="modal-content">
-                    <div class="modal-header">
-                        <h5 class="modal-title" id="exampleModalLabel"><i
-                                class="nav-icon fas fa-calendar-alt my-1 btn-sm-1"></i> Tambah Tahun Pelajaran</h5>
-                        <button type="button" class="close" data-dismiss="modal" aria-label="Close">
-                            <span aria-hidden="true">&times;</span>
-                        </button>
-                    </div>
-                    <div class="modal-body">
-                        <form action="/tapel/tambah" method="POST">
-                            {{csrf_field()}}
-                            <div class="row">
+            <div class="modal fade" id="tambahTapel" tabindex="-1" role="dialog" aria-labelledby="exampleModalLabel" aria-hidden="true">
+                <div class="modal-dialog" role="document">
+                    <div class="modal-content">
+                        <div class="modal-header">
+                            <h5 class="modal-title" id="exampleModalLabel"><i class="nav-icon fas fa-calendar-alt my-1 btn-sm-1"></i> Tambah Tahun Pelajaran</h5>
+                            <button type="button" class="close" data-dismiss="modal" aria-label="Close">
+                                <span aria-hidden="true">&times;</span>
+                            </button>
+                        </div>
+                        <div class="modal-body">
+                            <form action="/tapel/tambah" method="POST">
+                                {{csrf_field()}}
+                                <div class="row">
                                     <label for="tahun">Tahun Pelajaran</label>
-                                    <input value="{{old('tahun')}}" name="tahun" type="text" class="form-control bg-light"
-                                        id="tahun" placeholder="Contoh : (2019/2020)" required>
+                                    <input value="{{old('tahun')}}" name="tahun" type="text" class="form-control bg-light" id="tahun" placeholder="Contoh : (2019/2020)" required>
                                     <label for="semester">Semester</label>
-                                    <select name="semester" class="custom-select my-1 mr-sm-2 bg-light" id="semester"required>
+                                    <select name="semester" class="custom-select my-1 mr-sm-2 bg-light" id="semester" required>
                                         <option value="">-- Pilih Semester --</option>
                                         <option value="Semester Ganjil">Semester Ganjil</option>
                                         <option value="Semester Genap">Semester Genap</option>
                                     </select>
-                            </div>
-                            <hr>
-                            <button type="submit" class="btn btn-success btn-sm"><i class="fas fa-save"></i>
-                                SIMPAN</button>
-                        </form>
+                                </div>
+                                <hr>
+                                <button type="submit" class="btn btn-success btn-sm"><i class="fas fa-save"></i>
+                                    SIMPAN</button>
+                            </form>
+                        </div>
                     </div>
                 </div>
             </div>
-        </div>
         </div>
     </div>
 </section>
